Revert "Update to latest API version and libraries"
This reverts commit f7a358dc1ae003d07a7c325df1ceaf22da9f3954.

+++ app/src/main/java/paulscode/android/mupen64plusae/task/ExtractTexturesService.java
@@ -151,12 +151,14 @@ public class ExtractTexturesService extends Service
                         FileUtil.unzipAll( getApplicationContext(), mFileUri, outputFolder );
                     }
                 } else {
-                    headerName = TextureInfo.getTexturePackNameFromSevenZ(getApplicationContext(), mFileUri);
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                        headerName = TextureInfo.getTexturePackNameFromSevenZ(getApplicationContext(), mFileUri);
 
-                    if( !TextUtils.isEmpty( headerName ) ) {
-                        String outputFolder = globalPrefs.hiResTextureDir + headerName;
-                        FileUtil.deleteFolder( new File( outputFolder ) );
-                        FileUtil.unSevenZAll( getApplicationContext(), mFileUri, outputFolder );
+                        if( !TextUtils.isEmpty( headerName ) ) {
+                            String outputFolder = globalPrefs.hiResTextureDir + headerName;
+                            FileUtil.deleteFolder( new File( outputFolder ) );
+                            FileUtil.unSevenZAll( getApplicationContext(), mFileUri, outputFolder );
+                        }
                     }
                 }
 
@@ -187,6 +189,9 @@ public class ExtractTexturesService extends Service
     }
 
     public void initChannels(Context context) {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O) {
+            return;
+        }
 
         NotificationManager notificationManager =
                 (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
